working on jr jones
so far it's working on jr jones page but the way im finding reviews is by hardcoding how far down in the html it is, I need a way first to be able to transpose this point up or down, and second to find an automated way to set this transpose number. also error handling would be some icing on the cake.